Initial setup.
We build a docker image, extract its content, and call debootstick.

There is an issue with docker 1.6, we could not expose the filesystem
of a container through NFS. (Because of the use of the mount namespace,
making the mount hidden from the host.)
Because of this we use an older version, docker 1.3, from snapshot.debian.org.