Update docker package: use the one of debian backports.
Before, we were blocked to an old version of docker because
later versions did not allow to mount the filesystem root of
a running container.
It was a serious issue because these old versions of docker cannot
access the docker hub anymore.
In the python code (walt-python-packages project), we now mount
this root filesystem manually, without starting a container.
Thus we could upgrade docker.