Add ptp daemon and related changes.
ptp must start only when an ip is set on eth0.1, which
is done by walt-server during its init or platform setup
phase.
In order to achieve this, we instruct systemd that the ptp
daemon must be run after walt-server, and we turned the walt-server
systemd unit to the type "notify". This allows systemd to wait
for a notification from the walt-server process before considering
the process is started. Thus, the walt-server process first
performs its init or setup phase, then sends this notification,
which causes systemd to launch ptpd.

Note that since the init phase of walt-server may be long
(especially for the 1st run involving the platform setup), which
causes systemd to print warning messages on the console.
In order to avoid messing the walt server console screen with this
messages, we added a boot parameter 'console=tty1' in order to
direct these systemd messages elsewhere (walt server console is
using tty8).

There is also a matching that needs to be done to allow old
walt images to run on new servers and vice-versa. This is achieved
by introducing a file /etc/walt/server.conf on the server that defines
optional features implemented by this server. For now, only this
'ptp' feature is defined. New images provide a file /etc/walt/image.conf
which allows to match and enable only the features implemented on both
sides. If any of these 2 files is missing, we consider we use an old
server or image and thus no optional feature is possible.