rpi.uboot: get the optional dtb, then the kernel
In a WalT image, the dtb is optional, the kernel is not.
If we download the kernel first using tftp, and then an issue
occurs on server side, the dtb download failure might be
interpreted as if there was no dtb in the image. Then,
the kernel load will hang because of the NFS issue.
If we load the optional dtb first, and then the kernel,
any issue with the kernel download will cause the node
to reboot, and the risk of a NFS issue occuring between kernel
download and execution is minimized.